# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/role": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "查询角色列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色名",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.Response"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "查询用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户名",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/core.Response"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "data": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/definitions/core.PaginatedData"
                                                    },
                                                    {
                                                        "type": "object",
                                                        "properties": {
                                                            "list": {
                                                                "type": "array",
                                                                "items": {
                                                                    "$ref": "#/definitions/models.User"
                                                                }
                                                            }
                                                        }
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/{userID}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core.PaginatedData": {
            "type": "object",
            "properties": {
                "list": {},
                "total": {
                    "type": "integer"
                }
            }
        },
        "core.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.Organization": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "string"
                },
                "modifyTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "name": {
                    "description": "组织名",
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                }
            }
        },
        "models.Permission": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "string"
                },
                "modifyTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "name": {
                    "description": "权限名称",
                    "type": "string"
                },
                "parentID": {
                    "description": "父级权限",
                    "type": "string"
                }
            }
        },
        "models.Role": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "string"
                },
                "modifyTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 250
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Permission"
                    }
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "avatar": {
                    "description": "头像",
                    "type": "string",
                    "maxLength": 255
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "description": "主键",
                    "type": "string"
                },
                "modifyTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string",
                    "maxLength": 255
                },
                "organization": {
                    "description": "组织",
                    "$ref": "#/definitions/models.Organization"
                },
                "organizationID": {
                    "description": "组织ID",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "maxLength": 255
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "roles": {
                    "description": "角色",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Role"
                    }
                },
                "sex": {
                    "description": "性别 0: 男 1: 女",
                    "type": "integer"
                },
                "username": {
                    "description": "账户名",
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 0
                }
            }
        }
    }
}